Implement MultiPayloadEnum support for projectEnumValue
This code rearchitects and simplifies the projectEnumValue support by
introducing a new `TypeInfo` subclass for each kind of enum, including trivial,
no-payload, single-payload, and three different classes for multi-payload enums:

* "UnsupportedEnum" that we don't understand.  This will allow us to
  return correct "don't know" answers in cases where the
  runtime lacks enough information to accurately handle
  a particular enum.

* MP Enums that only use a separate tag value.  This includes
  generic enums and other dynamic layouts, as well as
  enums whose payloads have no spare bits.

* MP Enums that use spare bits, possibly in addition to
  a separate tag.  This logic can only be used, of course,
  if we can in fact compute a spare bit mask that agrees
  with the compiler.

The final challenge is to choose one of the above three
handlings for every MPE.  The current code that computes
the spare bit mask for the third type above has some flaws
that need to be understood.

Next step: Figure out spare bit mask computation.  Any cases we can compute
correctly based on the information available at runtime can be handled by one of
the above three TI classes.  For the remaining cases, we'll probably have to
divert to "Unsupported" until we can arrange for the compiler to provide
us with augmented data.