swift-syntax-parser-tool: tweak the build rules
Adjust the build rules to add a build RPATH into the executable to
ensure that `lib_InternalSwiftSyntaxParser.so` is properly found when
running the tests.  This also should ensure that we always use the
correct version, irrespective of the setting of `LD_LIBRARY_PATH`.  This
RPATH is marked as a `BUILD_RPATH` and will be stripped if the tool is
installed.

Adjust the dependencies to be clearer using generator expressions and
use the `target_compile_options` rather than the more fragile
`set_target_properties` to add `-fblocks` to the compilation.  This is
now possible as we are ensured that we are using CMake 3.15+.