create NOTICE as stated in license guidelines
According to [Apache License Document - Applying the license to new software](http://www.apache.org/dev/apply-license.html#new), it's required to put a NOTICE file in the root folder with correct years and owners of the software.

I would recommend `Swift` to follow the best practices and hence made this sample `NOTICE` file. Feel free to change and edit this as appropriate.