fix autocomplete crash (#30540)
A jupyter test case triggers this crash, where a method is called on ExprType when it is not null. This PR adds a null check and a test case in the compiler.

This code and crash occurs on master, so I will prepare a PR to master too.

I am not sure exactly what the logic should be when ExprType is null. Also there are other places in this function where ExprType is used without a null check, and maybe they should get a null check too. We could discuss this on the master PR with someone more knowledgeable about the code.