[build] Set language flag for C++ on OpenBSD.
See #12510 and SR-3635 for some context; OpenBSD exhibits the same
problem listed here. Without this, the build fails with an undefined
symbol error for __gnustep_objcxx_personality_v0. Per commentary on the
aforementioned PR, forcing the language to be interpreted as C++ solves
the problem.

There might be something different we can do here to diagnose and fix
this, but this is the simplest solution to get the build working and
enabling this for all C++ files should be tautologous and therefore
harmless.