shims: repair the build for android
The math `long double` variants are only on x86/x86_64 (FP128 is
currently unsupported so ARM64 support is excluded).  However, Windows
and Android are FP64 and should not define the `long double` variants.
The original condition was completely incorrect as the `double` variants
are available everywhere.  This repairs the builds on android ARM64.