Sema: Contextualize closures for function builder bodies also
The ContextualizeClosures walker re-parents and assigns discriminators to
autoclosures. We weren't doing this walk for function builder bodies,
which meant that proper invariants were not being established for the
implicit autoclosures synthesized for partially applied method references.

This is a recent regression from my change to build curry thunks for
unapplied method references in Sema.

Fixes <rdar://problem/61039516>.