Add copy_to_ref sil instruction
This sil instruction can be used to copy an address to a class reference. It does this by memcpying the address to the first element in the class reference (the element after the reference count object).

The source must contain all stored properties of the class reference and must be an address. See SIL.rst for more details.